+++ src/components/AppComponents/AppHeaderBurgerMenuSections/AppHeaderBurgerMenuSections.stories.tsx
@@ -1,5 +1,4 @@
-import React from "react";
-import { StoryFn, Meta } from "@storybook/react";
+import { Meta } from "@storybook/react";
 
 import Component from "./AppHeaderBurgerMenuSections";
 
@@ -9,9 +8,8 @@ export default {
   component: Component,
 } as Meta<typeof Component>;
 
-const Template: StoryFn<typeof Component> = (args) => <Component {...args} />;
-
-export const NewMenuLinks = Template.bind({});
-NewMenuLinks.args = {
-  burgerMenuSections: burgerMenuSections,
+export const NewMenuLinks = {
+  args: {
+    burgerMenuSections: burgerMenuSections,
+  },
 };