+++ src/components/GenericPagesComponents/GenericSummaryCardNavButton/GenericSummaryCardNavButton.stories.tsx
@@ -1,4 +1,4 @@
-import { StoryFn, Meta } from "@storybook/react";
+import { Meta } from "@storybook/react";
 
 import Component from "./GenericSummaryCardNavButton";
 
@@ -7,14 +7,12 @@ export default {
   argTypes: {},
 } as Meta<typeof Component>;
 
-const Template: StoryFn<typeof Component> = (args) => <Component {...args} />;
-
-export const GenericSummaryCardNavButton = Template.bind({});
-
-GenericSummaryCardNavButton.args = {
-  buttons: [
-    { label: "Who we are", href: "/first-one" },
-    { label: "Board", href: "/second-one" },
-    { label: "Leadership", href: "/third-one" },
-  ],
+export const GenericSummaryCardNavButton = {
+  args: {
+    buttons: [
+      { label: "Who we are", href: "/first-one" },
+      { label: "Board", href: "/second-one" },
+      { label: "Leadership", href: "/third-one" },
+    ],
+  },
 };