+++ src/components/GenericPagesComponents/LandingPageQuote/LandingPageQuote.stories.tsx
@@ -1,5 +1,4 @@
-import React from "react";
-import { StoryFn, Meta } from "@storybook/react";
+import { Meta } from "@storybook/react";
 
 import { LandingPageQuote as Component } from "./LandingPageQuote";
 
@@ -8,10 +7,9 @@ export default {
   argTypes: {},
 } as Meta<typeof Component>;
 
-const Template: StoryFn<typeof Component> = (args) => <Component {...args} />;
-
-export const LandingPagesQuote = Template.bind({});
-LandingPagesQuote.args = {
-  text: "This is an inspiring quote",
-  attribution: "I Wroteit",
+export const LandingPagesQuote = {
+  args: {
+    text: "This is an inspiring quote",
+    attribution: "I Wroteit",
+  },
 };