+++ src/components/SharedComponents/Grid.deprecated/GridArea.deprecated.stories.tsx
@@ -1,4 +1,4 @@
-import { StoryFn, Meta } from "@storybook/react";
+import { Meta } from "@storybook/react";
 
 import Component from "./GridArea.deprecated";
 
@@ -24,6 +24,4 @@ export default {
   },
 } as Meta<typeof Component>;
 
-const Template: StoryFn<typeof Component> = (args) => <Component {...args} />;
-
-export const GridArea = Template.bind({});
+export const GridArea = {};