+++ src/components/SharedComponents/SearchForm/SearchForm.stories.tsx
@@ -1,5 +1,4 @@
-import React from "react";
-import { StoryFn, Meta } from "@storybook/react";
+import { Meta } from "@storybook/react";
 
 import Component from "./SearchForm";
 
@@ -10,9 +9,8 @@ export default {
   component: Component,
 } as Meta<typeof Component>;
 
-const Template: StoryFn<typeof Component> = (args) => <Component {...args} />;
-
-export const SearchInput = Template.bind({});
-SearchInput.args = {
-  placeholderText: "Search by keyword or topic",
+export const SearchInput = {
+  args: {
+    placeholderText: "Search by keyword or topic",
+  },
 };