+++ src/node-lib/curriculum-api-2023/queries/specialistLessonDownload/specialistLessonDownload.schema.ts
@@ -60,17 +60,14 @@ export const SpecialistLessonDownloadSchema = z.object({
     ),
     expired: z.boolean().nullable(),
     copyrightContent: copyrightContentSchema,
-    isDownloadRegionRestricted: z.boolean().optional(),
+    geoRestricted: z.boolean().nullable(),
+    loginRequired: z.boolean().nullable(),
   }),
 });
 
 export type SpecialistLessonDownloads = z.infer<
   typeof SpecialistLessonDownloadSchema
-> & {
-  lesson: {
-    isDownloadRegionRestricted: boolean;
-  };
-};
+>;
 export type SpecialistLessonDownloadsPageData = {
   curriculumData: SpecialistLessonDownloads;
 };